[AAE-10779] add stories and remove old component
