--- lib/core/src/lib/userinfo/userinfo.module.ts
@@ -1,46 +1,0 @@
-/*!
- * @license
- * Copyright 2019 Alfresco Software, Ltd.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-import { CommonModule } from '@angular/common';
-import { NgModule } from '@angular/core';
-import { TranslateModule } from '@ngx-translate/core';
-import { PipeModule } from '../pipes/pipe.module';
-import { UserInfoComponent } from './components/user-info.component';
-import { MatButtonModule } from '@angular/material/button';
-import { MatMenuModule } from '@angular/material/menu';
-import { MatTabsModule } from '@angular/material/tabs';
-import { MatCardModule } from '@angular/material/card';
-
-@NgModule({
-    imports: [
-        CommonModule,
-        MatButtonModule,
-        MatMenuModule,
-        MatTabsModule,
-        MatCardModule,
-        TranslateModule,
-        PipeModule
-    ],
-    declarations: [
-        UserInfoComponent
-    ],
-    exports: [
-        UserInfoComponent
-    ]
-})
-export class UserInfoModule {
-}
