+++ lib/core/src/public-api.ts
@@ -17,7 +17,6 @@
 
 export * from './lib/about/index';
 export * from './lib/viewer/index';
-export * from './lib/userinfo/index';
 export * from './lib/toolbar/index';
 export * from './lib/pagination/index';
 export * from './lib/login/index';
